Wrote some stuff into function and added an about page and a welcome message
