+++ app.yaml
@@ -1,3 +1,3 @@
 runtime: python39
 
-entrypoint: streamlit r0_✅_Handwritten_digit_recognizer.py --server.enableCORS false --browser.serverAddress 0.0.0.0 --browser.gatherUsageStats false --server.port $PORT
+entrypoint: streamlit 0_🤖_Handwritten_digit_recognizer.py --server.enableCORS false --browser.serverAddress 0.0.0.0 --browser.gatherUsageStats false --server.port $PORT
