+++ recources/__init__.py
@@ -1,0 +1,1 @@
+from .get_neural_network import get_neural_network
